services/model-server/requirements.txt
grpcio
grpcio-tools
onnxruntime
huggingface-hub
numpy
Pillow